roachtest: move jepsen tests to package test
Release note: None
