+++ pkg/cmd/roachtest/BUILD.bazel
@@ -50,7 +50,6 @@ go_library(
         "interleavedpartitioned.go",
         "inverted_index.go",
         "java_helpers.go",
-        "jepsen.go",
         "jobs.go",
         "kv.go",
         "kvbench.go",
