+++ pkg/cmd/roachtest/tests/BUILD.bazel
@@ -5,12 +5,15 @@ go_library(
     srcs = [
         "build_info.go",
         "disk_stall.go",
+        "jepsen.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/cmd/roachtest/tests",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/cmd/roachtest/cluster",
+        "//pkg/cmd/roachtest/option",
         "//pkg/cmd/roachtest/registry",
+        "//pkg/cmd/roachtest/spec",
         "//pkg/cmd/roachtest/test",
         "//pkg/server/serverpb",
         "//pkg/util/httputil",
